+++ src/vw/BundleAdjustment/ControlNetwork.h
@@ -38,7 +38,7 @@ namespace ba {
     std::string m_date_time, m_description, m_chooserName;
     double m_focalplane_x, m_focalplane_y;
     double m_ephemeris_time;
-    size_t m_image_id;
+    uint64 m_image_id;
     bool m_ignore, m_pixels_dominant;
 
   public:
@@ -61,7 +61,7 @@ namespace ba {
     ControlMeasureType m_type;
 
     /// Constructor
-    ControlMeasure( float, float, float, float, size_t,
+    ControlMeasure( float, float, float, float, uint64,
                     ControlMeasureType type = ControlMeasure::Automatic );
     ControlMeasure( ControlMeasureType type = ControlMeasure::Automatic );
     ControlMeasure( std::istream& f, ControlStorageFmt fmt) {
@@ -137,8 +137,8 @@ namespace ba {
 
     /// Setting/Reading the identifier for the image from which this
     /// control point was derived.
-    size_t image_id() const { return m_image_id; }
-    void set_image_id(size_t image_id) { m_image_id = image_id; }
+    uint64 image_id() const { return m_image_id; }
+    void set_image_id(uint64 image_id) { m_image_id = image_id; }
 
     /// Setting/Reading the description
     std::string description() const { return m_description; }
